+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to `laravel-collection-macros` will be documented in this fi
 - added support for Laravel 5.5, remove support for older versions
 - removed `dd` and `dump` macros
 
+## 2.6.0 - 2017-08-22
+
+- Added `tail`, `eachCons`, `sliceBefore` and `chunkBy`
+
+## 2.5.0 - 2017-08-03
+- Added customization options for `sectionBy`
+
 ## 2.4.0 - 2017-08-02
 - Added: `sectionBy`